krb5: Move soft_terminate_krb5_child to static
(cherry picked from commit c15bd3aeb3bdc0af23b69bf277c2177a69c92bc3)
